bugfix: fix incorrect data size

+++ thesis_project/src/main/java/com/mateuszkmita/thesis/external/controller/dto/page/PageDto.java
@@ -4,6 +4,6 @@ public record PageDto<T>(
         Iterable<T> data,
         int page,
         int numberOfPages,
-        int allDataSize
+        long allDataSize
 ) {
 }
